+++ packages/ui/button/src/Button.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Button, ButtonProps } from './Button'
+import { Button, type ButtonProps } from './Button'
 
 export default {
   title: 'Components/Button',
@@ -8,7 +8,8 @@ export default {
   argTypes: {
     prefix: { control: { type: 'text' } },
     suffix: { control: { type: 'text' } },
-    children: { control: { type: 'text' } }
+    children: { control: { type: 'text' } },
+    disabled: { control: { type: 'boolean' } }
   }
 }
 
@@ -17,7 +18,7 @@ export const Default = (args: ButtonProps) => (
     {...args}
     classNames={
       args.classNames ?? {
-        root: 'bg-gray-300 rounded-md border-1 border-gray-400/50 text-gray-700 font-medium hover:bg-gray-400 active:scale-97'
+        root: 'bg-neutral-300 rounded-md border-1 border-neutral-400/50 text-neutral-700 font-medium hover:bg-neutral-400 active:scale-97'
       }
     }
   >